[MISC] Move seqan3::views::zip to seqan3/utility/views/
