+++ include/seqan3/alignment/aligned_sequence/debug_stream_alignment.hpp
@@ -21,8 +21,8 @@
 #include <seqan3/core/detail/debug_stream_tuple.hpp>
 #include <seqan3/range/views/slice.hpp>
 #include <seqan3/range/views/to_char.hpp>
-#include <seqan3/range/views/zip.hpp>
 #include <seqan3/utility/tuple/concept.hpp>
+#include <seqan3/utility/views/zip.hpp>
 
 namespace seqan3::detail
 {
