+++ include/seqan3/alignment/pairwise/detail/type_traits.hpp
@@ -31,13 +31,13 @@
 #include <seqan3/core/configuration/configuration.hpp>
 #include <seqan3/core/detail/empty_type.hpp>
 #include <seqan3/core/detail/template_inspection.hpp>
-#include <seqan3/range/views/zip.hpp>
 #include <seqan3/utility/detail/bits_of.hpp>
 #include <seqan3/utility/simd/simd_traits.hpp>
 #include <seqan3/utility/simd/simd.hpp>
 #include <seqan3/utility/type_traits/function_traits.hpp>
 #include <seqan3/utility/type_traits/lazy_conditional.hpp>
 #include <seqan3/utility/views/chunk.hpp>
+#include <seqan3/utility/views/zip.hpp>
 
 namespace seqan3::detail
 {
