+++ include/seqan3/alignment/pairwise/policy/alignment_matrix_policy.hpp
@@ -18,8 +18,8 @@
 #include <seqan3/alignment/configuration/align_config_band.hpp>
 #include <seqan3/alignment/pairwise/detail/alignment_algorithm_state.hpp>
 #include <seqan3/range/views/slice.hpp>
-#include <seqan3/range/views/zip.hpp>
 #include <seqan3/utility/type_traits/basic.hpp>
+#include <seqan3/utility/views/zip.hpp>
 
 namespace seqan3::detail
 {
