+++ include/seqan3/alignment/pairwise/policy/simd_find_optimum_policy.hpp
@@ -18,8 +18,8 @@
 #include <seqan3/alignment/pairwise/detail/alignment_algorithm_state.hpp>
 #include <seqan3/alignment/pairwise/policy/find_optimum_policy.hpp>
 #include <seqan3/core/detail/empty_type.hpp>
-#include <seqan3/range/views/zip.hpp>
 #include <seqan3/utility/simd/concept.hpp>
+#include <seqan3/utility/views/zip.hpp>
 
 namespace seqan3::detail
 {
