+++ test/include/seqan3/test/performance/naive_minimiser_hash.hpp
@@ -20,9 +20,9 @@
 #include <seqan3/core/range/type_traits.hpp>
 #include <seqan3/range/views/complement.hpp>
 #include <seqan3/range/views/detail.hpp>
-#include <seqan3/range/views/zip.hpp>
 #include <seqan3/search/views/kmer_hash.hpp>
 #include <seqan3/search/views/minimiser.hpp>
+#include <seqan3/utility/views/zip.hpp>
 
 namespace seqan3::detail
 {
