+++ test/performance/alignment/edit_distance_unbanded_benchmark.cpp
@@ -15,11 +15,11 @@
 
 #include <seqan3/alignment/pairwise/align_pairwise.hpp>
 #include <seqan3/alphabet/nucleotide/dna4.hpp>
-#include <seqan3/range/views/zip.hpp>
 #include <seqan3/test/alignment/align_pairwise_edit_distance.hpp>
 #include <seqan3/test/performance/units.hpp>
 #include <seqan3/test/performance/sequence_generator.hpp>
 #include <seqan3/test/seqan2.hpp>
+#include <seqan3/utility/views/zip.hpp>
 
 #ifdef SEQAN3_HAS_SEQAN2
 #include <seqan/align.h>
