+++ test/performance/utility/simd/views/to_simd_benchmark.cpp
@@ -18,12 +18,12 @@
 #include <seqan3/alphabet/nucleotide/dna4.hpp>
 #include <seqan3/range/container/aligned_allocator.hpp>
 #include <seqan3/range/views/to.hpp>
-#include <seqan3/range/views/zip.hpp>
 #include <seqan3/test/performance/sequence_generator.hpp>
 #include <seqan3/utility/simd/concept.hpp>
 #include <seqan3/utility/simd/simd_traits.hpp>
 #include <seqan3/utility/simd/simd.hpp>
 #include <seqan3/utility/simd/views/to_simd.hpp>
+#include <seqan3/utility/views/zip.hpp>
 
 // ============================================================================
 //  naive implementation without condition inside of hot loop
