+++ test/snippet/alignment/scoring/nucleotide_scoring_scheme.cpp
@@ -4,7 +4,7 @@
 #include <seqan3/alphabet/nucleotide/dna5.hpp>
 #include <seqan3/alphabet/nucleotide/rna15.hpp>
 #include <seqan3/core/debug_stream.hpp>
-#include <seqan3/range/views/zip.hpp>
+#include <seqan3/utility/views/zip.hpp>
 
 int main()
 {
