+++ test/unit/alphabet/aminoacid/aa20_test.cpp
@@ -7,7 +7,7 @@
 
 #include <seqan3/alphabet/aminoacid/aa20.hpp>
 #include <seqan3/core/detail/debug_stream_range.hpp>
-#include <seqan3/range/views/zip.hpp>
+#include <seqan3/utility/views/zip.hpp>
 
 #include "../alphabet_constexpr_test_template.hpp"
 #include "../alphabet_test_template.hpp"
