+++ test/unit/core/algorithm/detail/execution_handler_template.hpp
@@ -16,10 +16,9 @@
 
 #include <seqan3/alphabet/nucleotide/dna4.hpp>
 #include <seqan3/range/views/type_reduce.hpp>
-#include <seqan3/range/views/zip.hpp>
-#include <seqan3/utility/views/chunk.hpp>
-
 #include <seqan3/test/performance/sequence_generator.hpp>
+#include <seqan3/utility/views/chunk.hpp>
+#include <seqan3/utility/views/zip.hpp>
 
 //Forward declaration
 namespace seqan3::detail
