+++ test/unit/io/detail/detail_record_test.cpp
@@ -12,9 +12,9 @@
 #include <seqan3/alphabet/nucleotide/dna4.hpp>
 #include <seqan3/alphabet/quality/phred42.hpp>
 #include <seqan3/io/detail/record.hpp>
-#include <seqan3/range/views/zip.hpp>
-
 #include <seqan3/test/expect_same_type.hpp>
+#include <seqan3/utility/views/zip.hpp>
+
 
 TEST(detail, select_types_with_ids)
 {
