+++ test/unit/io/sequence_file/sequence_file_format_test_template.hpp
@@ -19,9 +19,9 @@
 #include <seqan3/core/detail/debug_stream_alphabet.hpp>
 #include <seqan3/io/sequence_file/input.hpp>
 #include <seqan3/io/sequence_file/output.hpp>
-#include <seqan3/range/views/zip.hpp>
 #include <seqan3/test/expect_range_eq.hpp>
 #include <seqan3/test/pretty_printing.hpp>
+#include <seqan3/utility/views/zip.hpp>
 #include <seqan3/utility/views/convert.hpp>
 
 using seqan3::operator""_dna5;
