+++ test/unit/utility/views/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_subdirectories()
+
+seqan3_test (zip_test.cpp)
